+++ contracts/market/MarketSettings.sol
@@ -9,7 +9,7 @@ import "../interfaces/IMarketSettings.sol";
  * @dev Contract module which holds the setting params for all markets.
  */
 contract MarketSettings is IMarketSettings, Ownable, Initializable {
-    mapping(bytes32 => int256) private intVals;
+    mapping(bytes32 => int) private intVals;
 
     function initialize() external onlyInitializeOnce {
         _transferOwnership(msg.sender);
@@ -17,20 +17,17 @@ contract MarketSettings is IMarketSettings, Ownable, Initializable {
 
     /*=== setters ===*/
 
-    function setIntVals(bytes32 _key, int256 _value) external onlyOwner {
+    function setIntVals(bytes32 _key, int _value) external onlyOwner {
         intVals[_key] = _value;
     }
 
     /*=== getters ===*/
 
-    function getIntVals(bytes32 _key) external view returns (int256) {
+    function getIntVals(bytes32 _key) external view returns (int) {
         return intVals[_key];
     }
 
-    function getIntValsByMarket(
-        bytes32 _marketKey,
-        bytes32 _key
-    ) external view returns (int256) {
+    function getIntValsByMarket(bytes32 _marketKey, bytes32 _key) external view returns (int) {
         return intVals[keccak256(abi.encodePacked([_marketKey, _key]))];
     }
 }
